#213 Refactor shared to single page `/aukojimas`
