--- src/shared-pages/Aukojimas/index.css
@@ -1,15 +1,0 @@
-.Card {
-  &__actions {
-    display: flex;
-    flex-wrap: wrap;
-    margin: 20px 0;
-
-    @media (min-width: 41rem) {
-      flex-wrap: nowrap;
-    }
-  }
-}
-
-.TabsContainer {
-  display: flex;
-}
